+++ contracting_process/resource_level/coherent/milestone_status.py
@@ -2,19 +2,22 @@
 
 from pelican.util.checks import complete_result_resource, get_empty_result_resource
 from pelican.util.getter import deep_get, get_values
+from pelican.util.schema import get_paths
 
 version = 1.0
 applicable_statuses = {"scheduled", "notMet"}
+paths = get_paths("Milestone")
 
 
 def calculate(item):
     result = get_empty_result_resource(version)
 
-    milestones = []
-    for path in ("planning", "tender", "contracts", "contracts.implementation"):
-        milestones.extend(
-            v for v in get_values(item, f"{path}.milestones") if deep_get(v["value"], "status") in applicable_statuses
-        )
+    milestones = [
+        value
+        for path in paths
+        for value in get_values(item, path)
+        if deep_get(value["value"], "status") in applicable_statuses
+    ]
 
     if not milestones:
         result["meta"] = {"reason": "no milestone is unmet"}