+++ contracting_process/resource_level/coherent/release_date.py
@@ -8,44 +8,21 @@ Coherence check for release date.
 
 from pelican.util.checks import coherent_dates_check
 from pelican.util.getter import get_values
+from pelican.util.schema import get_paths
 
 version = 1.0
+paths = (
+    "awards.date",
+    "contracts.dateSigned",
+    "contracts.implementation.transactions.date",
+    *(f"{path}.date" for path in get_paths("Amendment")),
+    *(f"{path}.{key}" for path in get_paths("Milestone") for key in ("dateModified", "dateMet")),
+    *(f"{path}.{key}" for path in get_paths("Document") for key in ("datePublished", "dateModified")),
+)
 
 
 def calculate(item):
-    first_dates = []
-
-    for first_path in (
-        # amendments_dates
-        "tender.amendments.date",
-        "awards.amendments.date",
-        "contracts.amendments.date",
-        # dates
-        "awards.date",
-        "contracts.dateSigned",
-        "contracts.implementation.transactions.date",
-        # milestones_dates
-        "planning.milestones.dateModified",
-        "planning.milestones.dateMet",
-        "tender.milestones.dateModified",
-        "tender.milestones.dateMet",
-        "contracts.milestones.dateModified",
-        "contracts.milestones.dateMet",
-        "contracts.implementation.milestones.dateModified",
-        "contracts.implementation.milestones.dateMet",
-        # documents_dates
-        "planning.documents.datePublished",
-        "planning.documents.dateModified",
-        "tender.documents.datePublished",
-        "tender.documents.dateModified",
-        "awards.documents.datePublished",
-        "awards.documents.dateModified",
-        "contracts.documents.datePublished",
-        "contracts.documents.dateModified",
-        "contracts.implementation.documents.datePublished",
-        "contracts.implementation.documents.dateModified",
-    ):
-        first_dates.extend(get_values(item, first_path))
+    first_dates = [value for path in paths for value in get_values(item, path)]
 
     second_dates = get_values(item, "date")